Wifi Config APP start

+++ wifi_firmware_esp32/src/csma_control/contention_backoff.h
@@ -14,7 +14,7 @@ class CONTENTION_BACKOFF{
          /**
           * max value contention window allows
           */
-         uint8_t maximum;
+         uint16_t maximum;
          
     public:
